Исправляет замечания в 5 разделе
